Moved everything from Driver to MainController to enable resizing the window on game start.
