+++ src/Engine/MainController.java
@@ -3,6 +3,12 @@ package Engine;
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.Graphics;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
+
+import javax.swing.JFrame;
 
 import CommandingOfficers.CmdrStrong;
 import CommandingOfficers.Commander;
@@ -11,10 +17,12 @@ import UI.InputHandler;
 import UI.Art.SpriteArtist.SpriteMainMenuView;
 import UI.Art.SpriteArtist.SpriteMapView;
 
-public class MainController extends javax.swing.JPanel implements IController
+public class MainController extends javax.swing.JPanel implements IController, IView, ActionListener, KeyListener
 {
   private static final long serialVersionUID = 5548786952371603112L;
 
+  JFrame gameWindow;
+  
   private IController activeSubController;
   private IView activeSubView;
 
@@ -28,14 +36,27 @@ public class MainController extends javax.swing.JPanel implements IController
 
   private Dimension dimensions = new Dimension(240, 160);
 
+  private javax.swing.Timer repaintTimer;
+
   public MainController()
   {
     // Initialize member data.
     setPreferredSize(dimensions);
     
+    gameWindow = new JFrame();
+    gameWindow.add(this);
+    gameWindow.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+    gameWindow.addKeyListener(this);
+    gameWindow.pack();
+    gameWindow.setVisible(true);
+    
     // At game start, MainController will handle the inputs/rendering.
     activeSubView = new SpriteMainMenuView(this);
     activeSubController = null;
+    
+    // Draw the screen at (ideally) 60fps.
+    repaintTimer = new javax.swing.Timer(16, this);
+    repaintTimer.start();
   }
 
   public int getHighlightedOption()
@@ -84,7 +105,9 @@ public class MainController extends javax.swing.JPanel implements IController
 
             SpriteMapView smv = new SpriteMapView(newGame);
             activeSubView = smv;
-            setPreferredSize(activeSubView.getPreferredDimensions());
+            System.out.println("Setting dims to " + activeSubView.getPreferredDimensions().width + ", " + activeSubView.getPreferredDimensions().height);
+            gameWindow.getContentPane().setSize(activeSubView.getPreferredDimensions());
+            gameWindow.pack();
             MapController mapController = new MapController(newGame, smv);
 
             activeSubController = mapController;
@@ -126,11 +149,75 @@ public class MainController extends javax.swing.JPanel implements IController
   @Override // from JComponent
   public Dimension getPreferredSize()
   {
+    System.out.println("mc.gPS");
     if(null != activeSubView)
     {
       System.out.println("Size " + activeSubView.getPreferredDimensions().width + ", " + activeSubView.getPreferredDimensions().height);
       return activeSubView.getPreferredDimensions();
     }
     return dimensions;
+  }
+
+  @Override // From KeyListener
+  public void keyPressed(KeyEvent arg0)
+  {
+    boolean exitGame = false;
+
+    InputHandler.InputAction action = InputHandler.pressKey(arg0);
+
+    if( action != InputHandler.InputAction.NO_ACTION )
+    {
+      // Pass the action on to the main game controller. If it says it's done, close the program.
+      exitGame = handleInput(action);
+    }
+
+    if(exitGame)
+    {
+      // We are done here. I would say clean up data or whatever, but this is Java.
+      System.exit(0);
+    }
+  }
+
+  @Override // From KeyListener
+  public void keyReleased(KeyEvent arg0)
+  {
+    InputHandler.releaseKey(arg0);
+  }
+
+  @Override // From KeyListener
+  public void keyTyped(KeyEvent arg0)
+  {
+    // Don't care.
+  }
+
+  @Override // from ActionListener
+  public void actionPerformed(ActionEvent arg0)
+  {
+    // Redraw the screen if needed.
+    gameWindow.repaint();
+  }
+
+  @Override // From IView
+  public Dimension getPreferredDimensions()
+  {
+    return getPreferredSize();
+  }
+
+  @Override // From IView
+  public int getViewWidth()
+  {
+    return getWidth();
+  }
+
+  @Override // From IView
+  public int getViewHeight()
+  {
+    return getHeight();
+  }
+
+  @Override // From IView
+  public void render(Graphics g)
+  {
+    // Not used.
   }
 }
